Testing/Experiments/llamaindex/requirements.txt
llama-index
llama_hub
atlassian-python-api
weaviate-client
